refactor GUITests
