+++ src/test/java/org/cicirello/ibp/GUITestCases.java
@@ -1,6 +1,6 @@
 /*
  * Interactive Bin Packing.
- * Copyright (C) 2008, 2010, 2020-2022 Vincent A. Cicirello
+ * Copyright (C) 2008, 2010, 2020-2023 Vincent A. Cicirello
  *
  * This file is part of Interactive Bin Packing.
  *
@@ -25,23 +25,14 @@ import static org.junit.jupiter.api.Assertions.*;
 
 import java.awt.BorderLayout;
 import java.awt.Component;
-import java.awt.GraphicsEnvironment;
-import java.io.File;
-import java.io.FileReader;
-import java.io.IOException;
-import java.net.URISyntaxException;
-import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import javax.swing.JButton;
 import javax.swing.JComboBox;
 import javax.swing.JLabel;
-import javax.swing.JMenu;
-import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 import javax.swing.JTextArea;
 import javax.swing.JTextField;
 import org.junit.jupiter.api.*;
-import org.junit.jupiter.api.io.TempDir;
 
 /**
  * JUnit tests for all of the GUI classes of the Interactive Bin Packing Application. These test
@@ -51,572 +42,6 @@ import org.junit.jupiter.api.io.TempDir;
  * well as hyperlinks within the Tutorial and Help dialogs.
  */
 public class GUITestCases {
-
-  @Test
-  public void testInfoMenu() {
-    InfoDialog.setTestingMode();
-
-    int[] sizes = {7, 2, 18, 3, 6};
-    CallBack cb =
-        new CallBack() {
-          @Override
-          public void call() {}
-        };
-    Floor floor = new Floor(sizes);
-    ApplicationState state = new ApplicationState(1, floor, cb, cb, cb);
-    MenuBar menus = new MenuBar(null, state);
-    assertNull(menus.getTutorial());
-    assertNull(menus.getHelp());
-    JMenu infoMenu = menus.getMenu(4);
-
-    GraphicsEnvironment ge = GraphicsEnvironment.getLocalGraphicsEnvironment();
-    if (!ge.isHeadless()) {
-
-      infoMenu.getItem(0).doClick();
-      assertNotNull(menus.getTutorial());
-      assertEquals(1, menus.getTutorial().getActivationCount());
-
-      infoMenu.getItem(1).doClick();
-      assertNotNull(menus.getHelp());
-      assertEquals(1, menus.getHelp().getActivationCount());
-
-      infoMenu.getItem(0).doClick();
-      assertNotNull(menus.getTutorial());
-      assertEquals(2, menus.getTutorial().getActivationCount());
-
-      infoMenu.getItem(0).doClick();
-      assertNotNull(menus.getTutorial());
-      assertEquals(3, menus.getTutorial().getActivationCount());
-
-      infoMenu.getItem(1).doClick();
-      assertNotNull(menus.getHelp());
-      assertEquals(2, menus.getHelp().getActivationCount());
-
-      menus.getTutorial().dispose();
-      menus.getHelp().dispose();
-    }
-  }
-
-  @Test
-  public void testUIInit() {
-    UI ui = new UI(null);
-    ApplicationState state = ui.getApplicationState();
-    assertEquals(ApplicationState.MODE_PRACTICE, state.getMode());
-    assertEquals("Practice Mode", state.getModeString());
-    assertEquals("practice", state.getModeName());
-    Floor f = state.getFloor();
-    ArrayList<Item> initialFloorItems = f.getContents();
-    assertEquals(20, initialFloorItems.size());
-    ArrayList<Bin> bins = state.getBins();
-    assertEquals(9, bins.size());
-    for (Bin b : bins) {
-      assertEquals(100, b.capacity());
-      assertEquals(0, b.used());
-      assertTrue(b.isEmpty());
-    }
-    ArrayList<Item> items = state.getItems();
-    assertEquals(20, items.size());
-    for (Item i : items) {
-      assertTrue(f.contains(i));
-    }
-
-    int[] sizes2 = {15, 5, 25, 10, 20};
-    int[] sorted = {5, 10, 15, 20, 25};
-    Floor f2 = new Floor(sizes2);
-    state.setNewInstance(f2, "TestCase");
-    f = state.getFloor();
-    ArrayList<Item> floorItems = f.getContents();
-    assertEquals(sizes2.length, floorItems.size());
-    bins = state.getBins();
-    assertEquals(9, bins.size());
-    for (Bin b : bins) {
-      assertEquals(100, b.capacity());
-      assertEquals(0, b.used());
-      assertTrue(b.isEmpty());
-    }
-    items = state.getItems();
-    assertEquals(sizes2.length, items.size());
-    int j = 0;
-    for (Item i : items) {
-      assertTrue(f.contains(i));
-      assertEquals(sizes2[j], i.size());
-      j++;
-    }
-
-    state.sortFloor(false);
-    floorItems = f.getContents();
-    for (int i = 0; i < floorItems.size(); i++) {
-      assertEquals(sorted[i], floorItems.get(i).size());
-    }
-    state.sortFloor(true);
-    floorItems = f.getContents();
-    for (int i = 0; i < floorItems.size(); i++) {
-      assertEquals(sorted[sorted.length - i - 1], floorItems.get(i).size());
-    }
-  }
-
-  @Test
-  public void testModeMenu() {
-    int[] sizes = {7, 2, 18, 3, 6};
-    CallBack cb =
-        new CallBack() {
-          @Override
-          public void call() {}
-        };
-    Floor floor = new Floor(sizes);
-    ApplicationState state = new ApplicationState(1, floor, cb, cb, cb);
-    MenuBar menus = new MenuBar(null, state);
-    state.getFloor().remove(new Item("A", 7));
-    state.getBins().get(0).add(new Item("A", 7));
-    assertEquals(sizes.length - 1, state.getFloor().getContents().size());
-    assertEquals(1, state.getBins().get(0).getContents().size());
-    JMenu modeMenu = menus.getMenu(0);
-    assertTrue(modeMenu.getItem(0).isSelected());
-    for (int i = 1; i <= 4; i++) {
-      assertFalse(modeMenu.getItem(i).isSelected());
-    }
-    JMenu opsMenu = menus.getMenu(2);
-    assertTrue(opsMenu.getItem(0).isEnabled());
-    assertTrue(opsMenu.getItem(1).isEnabled());
-    modeMenu.getItem(1).doClick();
-    for (int i = 0; i <= 4; i++) {
-      if (i != 1) assertFalse(modeMenu.getItem(i).isSelected());
-      else assertTrue(modeMenu.getItem(i).isSelected());
-    }
-    assertFalse(opsMenu.getItem(0).isEnabled());
-    assertFalse(opsMenu.getItem(1).isEnabled());
-    assertEquals(sizes.length, state.getFloor().getContents().size());
-    assertEquals(0, state.getBins().get(0).getContents().size());
-    state.getFloor().remove(new Item("A", 7));
-    state.getBins().get(0).add(new Item("A", 7));
-    assertEquals(sizes.length - 1, state.getFloor().getContents().size());
-    assertEquals(1, state.getBins().get(0).getContents().size());
-
-    modeMenu.getItem(2).doClick();
-    for (int i = 0; i <= 4; i++) {
-      if (i != 2) assertFalse(modeMenu.getItem(i).isSelected());
-      else assertTrue(modeMenu.getItem(i).isSelected());
-    }
-    assertTrue(opsMenu.getItem(0).isEnabled());
-    assertTrue(opsMenu.getItem(1).isEnabled());
-    assertEquals(sizes.length, state.getFloor().getContents().size());
-    assertEquals(0, state.getBins().get(0).getContents().size());
-    state.getFloor().remove(new Item("A", 7));
-    state.getBins().get(0).add(new Item("A", 7));
-    assertEquals(sizes.length - 1, state.getFloor().getContents().size());
-    assertEquals(1, state.getBins().get(0).getContents().size());
-
-    modeMenu.getItem(3).doClick();
-    for (int i = 0; i <= 4; i++) {
-      if (i != 3) assertFalse(modeMenu.getItem(i).isSelected());
-      else assertTrue(modeMenu.getItem(i).isSelected());
-    }
-    assertFalse(opsMenu.getItem(0).isEnabled());
-    assertFalse(opsMenu.getItem(1).isEnabled());
-    assertEquals(sizes.length, state.getFloor().getContents().size());
-    assertEquals(0, state.getBins().get(0).getContents().size());
-    state.getFloor().remove(new Item("A", 7));
-    state.getBins().get(0).add(new Item("A", 7));
-    assertEquals(sizes.length - 1, state.getFloor().getContents().size());
-    assertEquals(1, state.getBins().get(0).getContents().size());
-
-    modeMenu.getItem(4).doClick();
-    for (int i = 0; i <= 4; i++) {
-      if (i != 4) assertFalse(modeMenu.getItem(i).isSelected());
-      else assertTrue(modeMenu.getItem(i).isSelected());
-    }
-    assertTrue(opsMenu.getItem(0).isEnabled());
-    assertTrue(opsMenu.getItem(1).isEnabled());
-    assertEquals(sizes.length, state.getFloor().getContents().size());
-    assertEquals(0, state.getBins().get(0).getContents().size());
-    state.getFloor().remove(new Item("A", 7));
-    state.getBins().get(0).add(new Item("A", 7));
-    assertEquals(sizes.length - 1, state.getFloor().getContents().size());
-    assertEquals(1, state.getBins().get(0).getContents().size());
-
-    modeMenu.getItem(3).doClick();
-    assertFalse(opsMenu.getItem(0).isEnabled());
-    assertFalse(opsMenu.getItem(1).isEnabled());
-    assertEquals(sizes.length, state.getFloor().getContents().size());
-    assertEquals(0, state.getBins().get(0).getContents().size());
-    state.getFloor().remove(new Item("A", 7));
-    state.getBins().get(0).add(new Item("A", 7));
-    assertEquals(sizes.length - 1, state.getFloor().getContents().size());
-    assertEquals(1, state.getBins().get(0).getContents().size());
-
-    modeMenu.getItem(0).doClick();
-    assertTrue(opsMenu.getItem(0).isEnabled());
-    assertTrue(opsMenu.getItem(1).isEnabled());
-    assertEquals(sizes.length - 1, state.getFloor().getContents().size());
-    assertEquals(1, state.getBins().get(0).getContents().size());
-  }
-
-  @Test
-  public void testOperationsMenu() {
-    int[] sizes = {7, 2, 18, 3, 6};
-    Item[] sorted = {
-      new Item("B", 2), new Item("D", 3), new Item("E", 6), new Item("A", 7), new Item("C", 18)
-    };
-    Item[] sorted2 = {new Item("B", 2), new Item("D", 3), new Item("E", 6), new Item("C", 18)};
-    CallBack cb =
-        new CallBack() {
-          @Override
-          public void call() {}
-        };
-    Floor floor = new Floor(sizes);
-    ApplicationState state = new ApplicationState(1, floor, cb, cb, cb);
-    MenuBar menus = new MenuBar(null, state);
-    JMenu opsMenu = menus.getMenu(2);
-    assertTrue(opsMenu.getItem(0).isEnabled());
-    assertTrue(opsMenu.getItem(1).isEnabled());
-    assertTrue(opsMenu.getItem(2).isEnabled());
-    opsMenu.getItem(1).doClick();
-    ArrayList<Item> shouldBeSorted = state.getFloor().getContents();
-    for (int i = 0; i < sorted.length; i++) {
-      assertEquals(sorted[i], shouldBeSorted.get(i));
-    }
-
-    floor = new Floor(sizes);
-    state = new ApplicationState(1, floor, cb, cb, cb);
-    menus = new MenuBar(null, state);
-    opsMenu = menus.getMenu(2);
-    assertTrue(opsMenu.getItem(0).isEnabled());
-    assertTrue(opsMenu.getItem(1).isEnabled());
-    assertTrue(opsMenu.getItem(2).isEnabled());
-    opsMenu.getItem(0).doClick();
-    shouldBeSorted = state.getFloor().getContents();
-    reverse(sorted);
-    for (int i = 0; i < sorted.length; i++) {
-      assertEquals(sorted[i], shouldBeSorted.get(i));
-    }
-    reverse(sorted);
-
-    floor = new Floor(sizes);
-    state = new ApplicationState(1, floor, cb, cb, cb);
-    menus = new MenuBar(null, state);
-    opsMenu = menus.getMenu(2);
-    assertTrue(opsMenu.getItem(0).isEnabled());
-    assertTrue(opsMenu.getItem(1).isEnabled());
-    assertTrue(opsMenu.getItem(2).isEnabled());
-    state.getFloor().remove(new Item("A", 7));
-    state.getBins().get(0).add(new Item("A", 7));
-    opsMenu.getItem(1).doClick();
-    shouldBeSorted = state.getFloor().getContents();
-    for (int i = 0; i < sorted2.length; i++) {
-      assertEquals(sorted2[i], shouldBeSorted.get(i));
-    }
-    assertTrue(state.getBins().get(0).contains(new Item("A", 7)));
-
-    floor = new Floor(sizes);
-    state = new ApplicationState(1, floor, cb, cb, cb);
-    menus = new MenuBar(null, state);
-    opsMenu = menus.getMenu(2);
-    assertTrue(opsMenu.getItem(0).isEnabled());
-    assertTrue(opsMenu.getItem(1).isEnabled());
-    assertTrue(opsMenu.getItem(2).isEnabled());
-    state.getFloor().remove(new Item("A", 7));
-    state.getBins().get(0).add(new Item("A", 7));
-    opsMenu.getItem(0).doClick();
-    shouldBeSorted = state.getFloor().getContents();
-    reverse(sorted2);
-    for (int i = 0; i < sorted2.length; i++) {
-      assertEquals(sorted2[i], shouldBeSorted.get(i));
-    }
-    assertTrue(state.getBins().get(0).contains(new Item("A", 7)));
-    reverse(sorted2);
-
-    class MenuBarTester extends MenuBar {
-      boolean lowerBoundButtonClicked;
-
-      MenuBarTester(InteractiveBinPacking f, ApplicationState state) {
-        super(f, state);
-        lowerBoundButtonClicked = false;
-      }
-
-      @Override
-      void displayLowerBoundMessage(String message) {
-        lowerBoundButtonClicked = true;
-      }
-    }
-    floor = new Floor(sizes);
-    state = new ApplicationState(1, floor, cb, cb, cb);
-    menus = new MenuBarTester(null, state);
-    opsMenu = menus.getMenu(2);
-    assertTrue(opsMenu.getItem(0).isEnabled());
-    assertTrue(opsMenu.getItem(1).isEnabled());
-    assertTrue(opsMenu.getItem(2).isEnabled());
-    assertFalse(((MenuBarTester) menus).lowerBoundButtonClicked);
-    opsMenu.getItem(2).doClick();
-    assertTrue(((MenuBarTester) menus).lowerBoundButtonClicked);
-  }
-
-  @Test
-  public void testSessionMenuSaveSession(@TempDir File sub) {
-    class MenuBarTester extends MenuBar {
-      int errorCount;
-      int confirmCount;
-      int confirmSaveResponse;
-
-      MenuBarTester(InteractiveBinPacking f, ApplicationState state) {
-        super(f, state);
-        errorCount = confirmCount = 0;
-      }
-
-      @Override
-      void displayErrorMessage(String message) {
-        errorCount++;
-      }
-
-      @Override
-      int confirmSave() {
-        confirmCount++;
-        return confirmSaveResponse;
-      }
-
-      void setResponse(int response) {
-        confirmSaveResponse = response;
-      }
-    }
-    int[] sizes = {7, 2, 18, 3, 6};
-    CallBack cb =
-        new CallBack() {
-          @Override
-          public void call() {}
-        };
-    Floor floor = new Floor(sizes);
-    ApplicationState state = new ApplicationState(1, floor, cb, cb, cb);
-    MenuBarTester menus = new MenuBarTester(null, state);
-
-    /* // OLD JUnit 4... JUnit 5 approach has this as a field annotated.
-    File sub = tempFolder.newFolder("sub");
-    */
-    File logFile = new File(sub, "testlog.ibp");
-    menus.saveSessionLog(logFile);
-    int index = 0;
-    try (FileReader in = new FileReader(logFile, StandardCharsets.UTF_8)) {
-      String session = state.loadSessionLog(in);
-      index = session.indexOf("SAVE_SESSION_LOG");
-      assertTrue(index >= 0);
-    } catch (IOException ex) {
-      fail();
-    }
-    assertEquals(0, menus.errorCount);
-    assertEquals(0, menus.confirmCount);
-
-    File without = new File(sub, "without.txt");
-    File with = new File(sub, "without.txt.ibp");
-    menus.saveSessionLog(without);
-    try (FileReader in = new FileReader(with, StandardCharsets.UTF_8)) {
-      String session = state.loadSessionLog(in);
-      index++;
-      index = session.indexOf("LOAD_SESSION_LOG", index);
-      assertTrue(index >= 0);
-      index++;
-      index = session.indexOf("SAVE_SESSION_LOG", index);
-      assertTrue(index >= 0);
-    } catch (IOException ex) {
-      fail();
-    }
-    assertEquals(0, menus.errorCount);
-    assertEquals(0, menus.confirmCount);
-
-    menus.setResponse(JOptionPane.YES_OPTION);
-    menus.saveSessionLog(logFile);
-    try (FileReader in = new FileReader(logFile, StandardCharsets.UTF_8)) {
-      String session = state.loadSessionLog(in);
-      index++;
-      index = session.indexOf("LOAD_SESSION_LOG", index);
-      assertTrue(index >= 0);
-      index++;
-      index = session.indexOf("SAVE_SESSION_LOG", index);
-      assertTrue(index >= 0);
-    } catch (IOException ex) {
-      fail();
-    }
-    assertEquals(0, menus.errorCount);
-    assertEquals(1, menus.confirmCount);
-
-    menus.setResponse(JOptionPane.NO_OPTION);
-    menus.saveSessionLog(logFile);
-    try (FileReader in = new FileReader(logFile, StandardCharsets.UTF_8)) {
-      String session = state.loadSessionLog(in);
-      index++;
-      int tempIndex = session.indexOf("LOAD_SESSION_LOG", index);
-      assertTrue(tempIndex < 0);
-      tempIndex = session.indexOf("SAVE_SESSION_LOG", index);
-      assertTrue(tempIndex < 0);
-    } catch (IOException ex) {
-      fail();
-    }
-    assertEquals(0, menus.errorCount);
-    assertEquals(2, menus.confirmCount);
-
-    try {
-      menus.setResponse(JOptionPane.YES_OPTION);
-      File dir = new File(InteractiveBinPacking.class.getResource("testlogs/dir.ibp").toURI());
-      menus.saveSessionLog(dir);
-      assertEquals(1, menus.errorCount);
-      assertEquals(3, menus.confirmCount);
-    } catch (URISyntaxException ex) {
-      fail();
-    }
-  }
-
-  @Test
-  public void testSessionMenuLoadSession() {
-    class MenuBarTester extends MenuBar {
-      int errorCount;
-
-      MenuBarTester(InteractiveBinPacking f, ApplicationState state) {
-        super(f, state);
-        errorCount = 0;
-      }
-
-      @Override
-      void displayErrorMessage(String message) {
-        errorCount++;
-      }
-    }
-    int[] sizes = {7, 2, 18, 3, 6};
-    CallBack cb =
-        new CallBack() {
-          @Override
-          public void call() {}
-        };
-    Floor floor = new Floor(sizes);
-    ApplicationState state = new ApplicationState(1, floor, cb, cb, cb);
-    MenuBarTester menus = new MenuBarTester(null, state);
-    File nonExistent = new File("ThisFileDoesNotExist.ibp");
-    menus.loadSessionLog(nonExistent);
-    assertEquals(1, menus.errorCount);
-    try {
-      File wrongType = new File(InteractiveBinPacking.class.getResource("html/about.html").toURI());
-      menus.loadSessionLog(wrongType);
-      assertEquals(2, menus.errorCount);
-    } catch (URISyntaxException ex) {
-      fail();
-    }
-    try {
-      File malformed =
-          new File(InteractiveBinPacking.class.getResource("testlogs/malformed.ibp").toURI());
-      menus.loadSessionLog(malformed);
-      assertEquals(3, menus.errorCount);
-    } catch (URISyntaxException ex) {
-      fail();
-    }
-    try {
-      File dir = new File(InteractiveBinPacking.class.getResource("testlogs/dir.ibp").toURI());
-      menus.loadSessionLog(dir);
-      assertEquals(4, menus.errorCount);
-    } catch (URISyntaxException ex) {
-      fail();
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  @Test
-  public void testProblemMenu() {
-    int[] sizes = {7, 2, 18, 3, 6};
-    CallBack cb =
-        new CallBack() {
-          @Override
-          public void call() {}
-        };
-    Floor floor = new Floor(sizes);
-    ApplicationState state = new ApplicationState(1, floor, cb, cb, cb);
-    MenuBar menus = new MenuBar(null, state);
-    state.getFloor().remove(new Item("A", 7));
-    state.getBins().get(0).add(new Item("A", 7));
-    assertEquals(sizes.length - 1, state.getFloor().getContents().size());
-    assertEquals(1, state.getBins().get(0).getContents().size());
-    JMenu problemMenu = menus.getMenu(1);
-    problemMenu.getItem(0).doClick();
-    ArrayList<Item> onFloor = state.getFloor().getContents();
-    assertEquals(20, onFloor.size());
-    assertEquals(0, state.getBins().get(0).getContents().size());
-    int[] expectedSizes = {
-      36, 33, 39, 43, 7, 19, 37, 8, 29, 28, 37, 23, 29, 10, 22, 11, 33, 9, 17, 30
-    };
-    char c = 'A';
-    for (int i = 0; i < expectedSizes.length; i++, c++) {
-      Item expected = new Item(c + "", expectedSizes[i]);
-      assertEquals(expected, onFloor.get(i));
-    }
-
-    floor = new Floor(sizes);
-    state = new ApplicationState(1, floor, cb, cb, cb);
-    menus = new MenuBar(null, state);
-    state.getFloor().remove(new Item("A", 7));
-    state.getBins().get(0).add(new Item("A", 7));
-    assertEquals(sizes.length - 1, state.getFloor().getContents().size());
-    assertEquals(1, state.getBins().get(0).getContents().size());
-    problemMenu = menus.getMenu(1);
-    problemMenu.getItem(1).doClick();
-    onFloor = state.getFloor().getContents();
-    assertEquals(20, onFloor.size());
-    assertEquals(0, state.getBins().get(0).getContents().size());
-    c = 'A';
-    boolean different = false;
-    for (int i = 0; i < expectedSizes.length && !different; i++, c++) {
-      Item old = new Item(c + "", expectedSizes[i]);
-      if (!old.equals(onFloor.get(i))) {
-        different = true;
-        break;
-      }
-    }
-    assertTrue(different);
-
-    class MenuBarTester extends MenuBar {
-
-      String[] values = {"hello", "world", "12", null};
-      int next;
-
-      MenuBarTester(InteractiveBinPacking f, ApplicationState state) {
-        super(f, state);
-      }
-
-      @Override
-      String getProblemInstanceNumberFromUser() {
-        String v = values[next];
-        next++;
-        return v;
-      }
-    }
-
-    ArrayList<Item> oldFloor = (ArrayList<Item>) onFloor.clone();
-    floor = new Floor(sizes);
-    state = new ApplicationState(1, floor, cb, cb, cb);
-    menus = new MenuBarTester(null, state);
-    state.getFloor().remove(new Item("A", 7));
-    state.getBins().get(0).add(new Item("A", 7));
-    assertEquals(sizes.length - 1, state.getFloor().getContents().size());
-    assertEquals(1, state.getBins().get(0).getContents().size());
-    problemMenu = menus.getMenu(1);
-    problemMenu.getItem(2).doClick();
-    onFloor = state.getFloor().getContents();
-    assertEquals(20, onFloor.size());
-    assertEquals(0, state.getBins().get(0).getContents().size());
-    c = 'A';
-    different = false;
-    for (int i = 0; i < expectedSizes.length && !different; i++, c++) {
-      Item old = oldFloor.get(i);
-      if (!old.equals(onFloor.get(i))) {
-        different = true;
-        break;
-      }
-    }
-    assertTrue(different);
-
-    oldFloor = (ArrayList<Item>) onFloor.clone();
-    problemMenu.getItem(2).doClick();
-    onFloor = state.getFloor().getContents();
-    assertEquals(20, onFloor.size());
-    assertEquals(0, state.getBins().get(0).getContents().size());
-    c = 'A';
-    for (int i = 0; i < expectedSizes.length; i++, c++) {
-      assertEquals(oldFloor.get(i), onFloor.get(i));
-    }
-  }
 
   @Test
   public void testTopPanel() {
@@ -1210,14 +635,6 @@ public class GUITestCases {
     to.setSelectedIndex(0);
     mButton.doClick();
     assertFalse(bottom.didErrorMessage());
-  }
-
-  private void reverse(Item[] array) {
-    for (int i = 0, j = array.length - 1; i < j; i++, j--) {
-      Item temp = array[i];
-      array[i] = array[j];
-      array[j] = temp;
-    }
   }
 
   private static class BottomPanelTestVersion extends BottomPanel {
